+++ Fabric/src/main/java/fuzs/stylisheffects/api/client/event/MobEffectWidgetEvents.java
@@ -28,6 +28,7 @@ public class MobEffectWidgetEvents {
         }
     });
 
+    @FunctionalInterface
     public interface MouseClicked {
 
         /**