+++ Forge/src/main/java/fuzs/stylisheffects/client/StylishEffectsForgeClient.java
@@ -3,12 +3,14 @@ package fuzs.stylisheffects.client;
 import fuzs.puzzleslib.client.core.ClientCoreServices;
 import fuzs.stylisheffects.StylishEffects;
 import fuzs.stylisheffects.client.handler.EffectScreenHandlerImpl;
+import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ContainerScreenEvent;
 import net.minecraftforge.client.event.RenderGuiOverlayEvent;
 import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.client.gui.overlay.VanillaGuiOverlay;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
@@ -23,11 +25,17 @@ public class StylishEffectsForgeClient {
     }
 
     private static void registerHandlers() {
+        MinecraftForge.EVENT_BUS.addListener((final TickEvent.ClientTickEvent evt) -> {
+            if (evt.phase == TickEvent.Phase.END) EffectScreenHandlerImpl.INSTANCE.onClientTick(Minecraft.getInstance());
+        });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.Opening evt) -> {
             EffectScreenHandlerImpl.INSTANCE.onScreenOpen(evt.getNewScreen());
         });
         MinecraftForge.EVENT_BUS.addListener((final ContainerScreenEvent.Render.Background evt) -> {
             EffectScreenHandlerImpl.INSTANCE.onDrawBackground(evt.getContainerScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY());
+        });
+        MinecraftForge.EVENT_BUS.addListener((final ContainerScreenEvent.Render.Foreground evt) -> {
+            EffectScreenHandlerImpl.INSTANCE.onDrawForeground(evt.getContainerScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY());
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.RenderInventoryMobEffects evt) -> {
             // disable vanilla effect rendering in inventory screen
@@ -38,9 +46,6 @@ public class StylishEffectsForgeClient {
                 EffectScreenHandlerImpl.INSTANCE.onRenderMobEffectIconsOverlay(evt.getPoseStack(), evt.getWindow().getGuiScaledWidth(), evt.getWindow().getGuiScaledHeight());
                 evt.setCanceled(true);
             }
-        });
-        MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.Init.Post evt) -> {
-            EffectScreenHandlerImpl.INSTANCE.onScreenInit(evt.getScreen());
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.MouseButtonPressed.Pre evt) -> {
             EffectScreenHandlerImpl.INSTANCE.onMouseClicked(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getButton());